refactor(website): move convert to ui code out to own file
